beginning of the recipe reshuffle, plus fixing up chat_with_tj
