+++ recipes/chat_with_tj/ibm-credentials.sample.env
