+++ recipes/simon_says/package.json
